Move stdlib catalog into keel-catalog and break types→runtime cycle
Relocate the per-namespace SPEC descriptor tables, the catalog()
aggregation, and the @tools capability metadata out of
runtime::namespaces and into the neutral keel-catalog crate. The type
checker (types::prelude, types::checker) and the interpreter capability
gate now read the stdlib surface from keel_catalog instead of reaching
into the runtime, so types no longer depends on runtime at all.

The runtime keeps the executable namespace() implementations and
cross-checks them against keel_catalog::specs::<ns>::SPEC in tests; the
catalog-integrity test moves to keel-catalog alongside the data.

Part of the workspace split (#71); unblocks extracting keel-compiler.
