+++ src/catalog.rs
@@ -11,12 +11,12 @@
 //! }
 //! ```
 
-pub use crate::builtins::{BuiltinMethod, BuiltinParam, BuiltinResult, TySpec};
+pub use keel_catalog::builtins::{BuiltinMethod, BuiltinParam, BuiltinResult, TySpec};
 
 /// Iterate every built-in namespace method registered in the stdlib catalog.
 ///
 /// The returned iterator visits each method exactly once in an unspecified
 /// but stable order within a single binary build.
 pub fn catalog() -> impl Iterator<Item = &'static BuiltinMethod> {
-    crate::types::prelude::catalog()
+    keel_catalog::catalog()
 }
