+++ src/runtime/namespaces/cache.rs
@@ -1,63 +1,7 @@
-use crate::builtins::{BuiltinMethod, BuiltinParam, BuiltinResult, TySpec};
 use crate::interpreter::Namespace;
 use crate::interpreter::value::Value;
 use crate::runtime::args::{expect_duration_named, expect_str};
 use crate::runtime::namespace::{ns, positional};
-
-pub(crate) const SPEC: &[BuiltinMethod] = &[
-    BuiltinMethod {
-        namespace: "cache",
-        name: "set",
-        params: &[
-            BuiltinParam {
-                name: "key",
-                ty: TySpec::Str,
-                optional: false,
-            },
-            BuiltinParam {
-                name: "value",
-                ty: TySpec::Dynamic,
-                optional: false,
-            },
-            BuiltinParam {
-                name: "ttl",
-                ty: TySpec::Duration,
-                optional: true,
-            },
-        ],
-        result: BuiltinResult::Fixed(TySpec::None_),
-        doc: "Store a value in the in-process cache, with an optional TTL duration.",
-    },
-    BuiltinMethod {
-        namespace: "cache",
-        name: "get",
-        params: &[BuiltinParam {
-            name: "key",
-            ty: TySpec::Str,
-            optional: false,
-        }],
-        result: BuiltinResult::Fixed(TySpec::NullableDynamic),
-        doc: "Retrieve a cached value, returning none if absent.",
-    },
-    BuiltinMethod {
-        namespace: "cache",
-        name: "delete",
-        params: &[BuiltinParam {
-            name: "key",
-            ty: TySpec::Str,
-            optional: false,
-        }],
-        result: BuiltinResult::Fixed(TySpec::None_),
-        doc: "Remove a key from the cache.",
-    },
-    BuiltinMethod {
-        namespace: "cache",
-        name: "clear",
-        params: &[],
-        result: BuiltinResult::Fixed(TySpec::None_),
-        doc: "Clear all entries from the cache.",
-    },
-];
 
 pub(crate) fn namespace() -> Namespace {
     ns!("cache", {
