+++ src/runtime/namespaces/log.rs
@@ -1,74 +1,8 @@
-use crate::builtins::{BuiltinMethod, BuiltinParam, BuiltinResult, TySpec};
 use crate::interpreter::value::Value;
 use crate::interpreter::{Host, Namespace};
 use crate::runtime::args::expect_str;
 use crate::runtime::context;
 use crate::runtime::namespace::{ns, positional};
-
-pub(crate) const SPEC: &[BuiltinMethod] = &[
-    BuiltinMethod {
-        namespace: "log",
-        name: "info",
-        params: &[BuiltinParam {
-            name: "message",
-            ty: TySpec::Dynamic,
-            optional: false,
-        }],
-        result: BuiltinResult::Fixed(TySpec::None_),
-        doc: "Emit an info-level log message.",
-    },
-    BuiltinMethod {
-        namespace: "log",
-        name: "warn",
-        params: &[BuiltinParam {
-            name: "message",
-            ty: TySpec::Dynamic,
-            optional: false,
-        }],
-        result: BuiltinResult::Fixed(TySpec::None_),
-        doc: "Emit a warning-level log message.",
-    },
-    BuiltinMethod {
-        namespace: "log",
-        name: "error",
-        params: &[BuiltinParam {
-            name: "message",
-            ty: TySpec::Dynamic,
-            optional: false,
-        }],
-        result: BuiltinResult::Fixed(TySpec::None_),
-        doc: "Emit an error-level log message.",
-    },
-    BuiltinMethod {
-        namespace: "log",
-        name: "debug",
-        params: &[BuiltinParam {
-            name: "message",
-            ty: TySpec::Dynamic,
-            optional: false,
-        }],
-        result: BuiltinResult::Fixed(TySpec::None_),
-        doc: "Emit a debug-level log message.",
-    },
-    BuiltinMethod {
-        namespace: "log",
-        name: "set_level",
-        params: &[BuiltinParam {
-            name: "level",
-            ty: TySpec::Str,
-            optional: false,
-        }],
-        result: BuiltinResult::Fixed(TySpec::None_),
-        doc: "Set the minimum log level (\"debug\", \"info\", \"warn\", or \"error\").",
-    },
-    BuiltinMethod {
-        namespace: "log",
-        name: "level",
-        params: &[],
-        result: BuiltinResult::Fixed(TySpec::Str),
-        doc: "Return the current log level as a string.",
-    },
-];
 
 fn log_if_enabled(host: &dyn Host, level: &str, msg: &str) {
     let call_rank = context::log_level_rank(level).unwrap_or(1);
